First batch of changes.
  * This incorporates the first batch of comments by Pat and the Tomas' first batch (regarding the Backends)

	modified:       Backends/include/gambit/Backends/backend_types/AlterBBN.hpp
	modified:       Backends/include/gambit/Backends/backend_types/MontePythonLike.hpp
	modified:       Backends/include/gambit/Backends/backend_types/classy.hpp
	modified:       Backends/include/gambit/Backends/backend_types/plc.hpp
	modified:       Backends/include/gambit/Backends/frontends/AlterBBN_2_2.hpp
	modified:       Backends/include/gambit/Backends/frontends/plc_3_0.hpp
	modified:       Backends/include/gambit/Backends/python_function.hpp
	modified:       Backends/include/gambit/Backends/python_helpers.hpp
	modified:       Backends/patches/classy/create_SDSSDR7_fid.py
	deleted:        Backends/patches/darkages/1.2.0/DarkAges_for_GAMBIT.patch
	modified:       Backends/patches/montepythonlike/3.3.0/MPLike_patch_script.py
	modified:       Backends/src/backend_types/AlterBBN.cpp
	modified:       Backends/src/backend_types/classy.cpp
	modified:       Backends/src/frontends/AlterBBN_2_2.cpp
	modified:       Backends/src/frontends/DarkAges_1_2_0.cpp
	modified:       Backends/src/frontends/MontePythonLike_3_3_0.cpp
	modified:       Backends/src/frontends/classy_2_6_3.cpp
	modified:       Backends/src/frontends/classy_2_9_3.cpp
	modified:       Backends/src/frontends/classy_2_9_4.cpp
	modified:       Backends/src/frontends/classy_exo_2_7_2.cpp
	modified:       Backends/src/frontends/plc_3_0.cpp
	modified:       CosmoBit/include/gambit/CosmoBit/CosmoBit_rollcall.hpp#
	modified:       CosmoBit/src/BBN.cpp
	modified:       Utils/include/gambit/Utils/util_functions.hpp

+++ Backends/include/gambit/Backends/frontends/AlterBBN_2_2.hpp
@@ -34,7 +34,7 @@ LOAD_LIBRARY
 BE_FUNCTION(Init_cosmomodel, void, (AlterBBN_2_2::relicparam*), "Init_cosmomodel", "Init_cosmomodel")
 BE_FUNCTION(nucl_err, int, (AlterBBN_2_2::relicparam*,double*,double*), "nucl_err", "nucl_err")
 
-BE_CONV_FUNCTION(get_NNUC, int, (), "get_NNUC")
+BE_CONV_FUNCTION(get_NNUC, size_t, (), "get_NNUC")
 BE_CONV_FUNCTION(get_abund_map_AlterBBN, map_str_int, (), "get_abund_map_AlterBBN")
 BE_CONV_FUNCTION(fill_cosmomodel, void, (AlterBBN_2_2::relicparam*, map_str_dbl &), "Init_AlterBBN")
 BE_CONV_FUNCTION(call_nucl_err, int, (map_str_dbl&,double*,double*), "call_nucl_err")
